requirements.txt
openai==0.28
streamlit
fpdf
python-dotenv